requirements.txt
colorama==0.4.6
Flask==3.1.0
PyYAML==5.3
PyYAML==6.0.2
